requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
requests
faster-whisper
sounddevice
webrtcvad-wheels